move get-error-message util function to shared lib
