+++ apps/vbpay/app/sign-in/[[...sign-in]]/hooks/use-reset-password-form.ts
@@ -1,9 +1,10 @@
 import { useState } from "react";
-import { getErrorMessage } from "@/utils/get-error-message";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { confirmResetPassword } from "aws-amplify/auth";
 import { useForm } from "react-hook-form";
 import { toast } from "sonner";
+
+import { getErrorMessage } from "@workspace/ui/lib/get-error-message";
 
 import { useErrorDialog } from "@/hooks/use-error-dialog";
